Create script to load database
