+++ config/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "linkfree",
+  "engines": {
+    "node": "18.x"
+  },
+  "type": "module"
+}
