+++ models/Link.js
@@ -21,4 +21,4 @@ const LinkSchema = new mongoose.Schema({
 
 LinkSchema.index({ username: 1, url: 1 });
 
-module.exports = mongoose.models.Link || mongoose.model("Link", LinkSchema);
+export default mongoose.models.Link || mongoose.model("Link", LinkSchema);
