+++ models/LinkStats.js
@@ -26,5 +26,4 @@ const linkStatsSchema = new mongoose.Schema({
 
 linkStatsSchema.index({ username: 1, date: 1 });
 
-module.exports =
-  mongoose.models.LinkStats || mongoose.model("LinkStats", linkStatsSchema);
+export default  mongoose.models.LinkStats || mongoose.model("LinkStats", linkStatsSchema);
