+++ models/ProfileStats.js
@@ -22,6 +22,5 @@ const profileStatsSchema = new mongoose.Schema({
 
 profileStatsSchema.index({ username: 1, date: 1 });
 
-module.exports =
-  mongoose.models.ProfileStats ||
+export default mongoose.models.ProfileStats ||
   mongoose.model("ProfileStats", profileStatsSchema);
