+++ models/User.js
@@ -19,4 +19,4 @@ const UserSchema = new mongoose.Schema({
   },
 });
 
-module.exports = mongoose.models.User || mongoose.model("User", UserSchema);
+export default mongoose.models.User || mongoose.model("User", UserSchema);
